% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goi_heatmap.R
\name{mod_goi_heatmap}
\alias{mod_goi_heatmap}
\title{Genes of Interest Heatmap Module}
\usage{
mod_goi_heatmap(input, output, session, filtered_data_rv)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{filtered_data_rv}{reactiveValues with norm_counts, samples, species}
}
\description{
Upload gene list and visualize as heatmap using user-selected metadata columns.
}
